src/requirements.txt
altair==4.1.0
ansicolors==1.1.8
ansiwrap==0.8.4
anyconfig @ file:///home/conda/feedstock_root/build_artifacts/anyconfig_1619079056336/work
anyio==3.0.1
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
arrow @ file:///Users/runner/miniforge3/conda-bld/arrow_1619571875722/work
astor==0.8.1
async-generator==1.10
attrs==19.3.0
Babel==2.9.1
backcall==0.2.0
bandit==1.7.0
base58==2.1.0
binaryornot==0.4.4
black==21.5b1
black-nb==0.5.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1619593780156/work
certifi==2020.12.5
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1613414067184/work
cfgv==3.2.0
chainmap==1.0.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorclass==2.2.0
combomethod==1.0.12
cookiecutter==1.7.2
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851512278/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
detect-secrets==1.0.3
distlib==0.3.1
dynaconf @ file:///home/conda/feedstock_root/build_artifacts/dynaconf_1615236842834/work
entrypoints==0.3
fastapi==0.63.0
fastavro==0.22.13
feast==0.10.4
filelock==3.0.12
flake8==3.9.1
Flask==1.1.2
fsspec==0.8.7
future==0.18.2
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1616781027402/work
gitdb2==4.0.2
GitPython==3.0.6
google-api-core==1.26.3
google-auth==1.30.0
googleapis-common-protos==1.52.0
graphviz==0.16
greenlet==1.1.0
grimp==1.2.3
grpcio==1.37.1
hypothesis==6.10.1
identify==2.2.4
idna==2.10
import-linter==1.2.1
importlib-metadata==1.7.0
iniconfig==1.1.1
intspan==1.6.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1589369830981/work
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-packaging==0.9.2
jupyter-server==1.6.4
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
jupyterlab-widgets==1.0.0
kedro @ file:///home/conda/feedstock_root/build_artifacts/kedro_1619034225149/work
kedro-viz==3.11.0
kiwisolver==1.3.1
makefile2dot==1.0.2
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1610127533370/work
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mementos==1.3.1
mistune==0.8.4
mmh3==3.0.0
mypy==0.812
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.5.1
networkx==2.5.1
nferx-sdk @ git+https://github.com/lumenbiomics/sdk@b42ad3788f58347ccb853c378eb0b435fde4cbd0
nodeenv==1.6.0
notebook==6.0.3
nulltype==2.3.1
numpy==1.20.3
options==1.4.10
packaging==20.9
pandas==1.2.4
pandavro==1.5.2
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.1
pbr==5.6.0
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
Pint==0.17
pip-check==2.6
pip-chill==1.0.1
pip-date==1.0.3
pip-review==1.1.0
pip-tools==5.5.0
pluggy @ file:///Users/runner/miniforge3/conda-bld/pluggy_1610318099099/work
poyo==0.5.0
pre-commit==2.12.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.15.8
ptyprocess==0.6.0
py==1.10.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydantic==1.8.1
pydeck==0.6.2
pyflakes==2.3.1
Pygments==2.6.1
PyHive==0.6.4
pymongo==3.10.1
PyMySQL==1.0.2
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291493748/work
pytest==6.2.3
python-dateutil==2.8.1
python-json-logger==0.1.11
python-slugify @ file:///home/conda/feedstock_root/build_artifacts/python-slugify_1619833713638/work
pytz==2021.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
say==1.6.6
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
semver==2.13.0
Send2Trash==1.5.0
simplere==1.2.13
six==1.16.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.3.0
SQLAlchemy==1.4.15
starlette==0.13.6
stevedore==3.3.0
streamlit==0.81.0
tabulate==0.8.9
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
text-unidecode==1.3
textdata==2.4.1
textwrap3==0.9.2
threadpoolctl==2.1.0
thrift==0.13.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomlkit==0.7.0
toolz==0.11.1
toposort==1.6
tornado==6.0.4
tqdm==4.60.0
traitlets==4.3.3
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
Unidecode @ file:///home/conda/feedstock_root/build_artifacts/unidecode_1612533191922/work
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
zipp==3.1.0
